Rename package and add a match control class
