+++ instrumentation/lettuce-6.5/src/main/java/com/nr/lettuce65/instrumentation/NRHolder.java
@@ -4,7 +4,7 @@
  *  * SPDX-License-Identifier: Apache-2.0
  *
  */
-package com.nr.lettuce6.instrumentation;
+package com.nr.lettuce65.instrumentation;
 
 import com.newrelic.api.agent.DatastoreParameters;
 import com.newrelic.api.agent.NewRelic;
