+++ instrumentation/lettuce-6.5/src/main/java/com/nr/lettuce65/instrumentation/RedisDatastoreParameters.java
@@ -1,4 +1,4 @@
-package com.nr.lettuce6.instrumentation;
+package com.nr.lettuce65.instrumentation;
 
 import com.newrelic.api.agent.DatastoreParameters;
 import io.lettuce.core.RedisURI;
@@ -7,9 +7,8 @@ public class RedisDatastoreParameters {
     public static DatastoreParameters from(RedisURI uri, String operation) {
         DatastoreParameters params;
         if (uri != null) {
-
             params = DatastoreParameters.product("Redis").collection(null).operation(operation)
-                    .instance(uri.getHost(), uri.getPort()).noDatabaseName().build();
+                    .instance(uri.getHost(), uri.getPort()).databaseName(String.valueOf(uri.getDatabase())).build();
         } else {
             params = DatastoreParameters.product("Redis").collection(null).operation(operation).noInstance()
                     .noDatabaseName().noSlowQuery().build();
