+++ instrumentation/lettuce-6.5/src/main/java/io/lettuce/core/AbstractRedisAsyncCommands_Instrumentation.java
@@ -12,8 +12,8 @@ import com.newrelic.api.agent.Segment;
 import com.newrelic.api.agent.Trace;
 import com.newrelic.api.agent.weaver.Weave;
 import com.newrelic.api.agent.weaver.Weaver;
-import com.nr.lettuce6.instrumentation.NRBiConsumer;
-import com.nr.lettuce6.instrumentation.RedisDatastoreParameters;
+import com.nr.lettuce65.instrumentation.NRBiConsumer;
+import com.nr.lettuce65.instrumentation.RedisDatastoreParameters;
 import io.lettuce.core.api.StatefulConnection;
 import io.lettuce.core.protocol.AsyncCommand;
 import io.lettuce.core.protocol.ProtocolKeyword;
